// frontend/src/types/url-route.type.ts
export enum UrlRouteType {
    expense = '#/expense',
    income = '#/income',
    createExpense = '#/create-expense',
    createIncome = '#/create-income',
    editExpense = '#/edit-expense',
    editIncome = '#/edit-income',
    incomeAndExpenses = '#/income-and-expenses',
    createIncomeAndExpenses = '#/create-income-and-expenses',
    editIncomeAndExpenses = '#/edit-income-and-expenses',
    login = '#/login',
    signup = '#/signup',
    logout = '#/logout'
}